move and delete and rename things so they make sense
